refactor: omit duplicate function
