+++ ext/node/lib.rs
@@ -619,7 +619,7 @@ deno_core::extension!(deno_node,
     "internal/http2/util.ts",
     "internal/idna.ts",
     "internal/net.ts",
-    "internal/normalize_encoding.mjs",
+    "internal/normalize_encoding.ts",
     "internal/options.ts",
     "internal/primordials.mjs",
     "internal/process/per_thread.mjs",
