+++ tools/core_import_map.json
@@ -144,7 +144,7 @@
     "ext:deno_node/internal/http.ts": "../ext/node/polyfills/internal/http.ts",
     "ext:deno_node/internal/http2/util.ts": "../ext/node/polyfills/internal/http2/util.ts",
     "ext:deno_node/internal/net.ts": "../ext/node/polyfills/internal/net.ts",
-    "ext:deno_node/internal/normalize_encoding.mjs": "../ext/node/polyfills/internal/normalize_encoding.mjs",
+    "ext:deno_node/internal/normalize_encoding.ts": "../ext/node/polyfills/internal/normalize_encoding.ts",
     "ext:deno_node/internal/options.ts": "../ext/node/polyfills/internal/options.ts",
     "ext:deno_node/internal/primordials.mjs": "../ext/node/polyfills/internal/primordials.mjs",
     "ext:deno_node/internal/process/per_thread.mjs": "../ext/node/polyfills/internal/process/per_thread.mjs",
